eventual conectivity

+++ app/src/main/java/com/example/ecostyle/view/CheckoutFragment.kt
@@ -1,5 +1,7 @@
 package com.example.ecostyle.view
 
+import android.content.Context
+import android.net.ConnectivityManager
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -13,10 +15,8 @@ import com.example.ecostyle.adapter.CartAdapter
 import com.example.ecostyle.model.CartItem
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
-import android.content.Context
 import kotlinx.coroutines.*
 import kotlinx.coroutines.tasks.await
-
 
 class CheckoutFragment : Fragment() {
 
@@ -34,8 +34,11 @@ class CheckoutFragment : Fragment() {
     ): View? {
         val view = inflater.inflate(R.layout.fragment_checkout, container, false)
 
+        // Inicializar las vistas
         totalPriceTextView = view.findViewById(R.id.total_price_text_view)
         recyclerView = view.findViewById(R.id.recycler_view_cart_items)
+        checkoutButton = view.findViewById(R.id.checkout_button)
+
         recyclerView.layoutManager = LinearLayoutManager(context)
 
         cartAdapter = CartAdapter(emptyList(), { cartItem ->
@@ -45,14 +48,17 @@ class CheckoutFragment : Fragment() {
         })
         recyclerView.adapter = cartAdapter
 
-        checkoutButton = view.findViewById(R.id.checkout_button)
-        checkoutButton.isEnabled = false
+        // Comportamiento del botón de checkout
+        checkoutButton.setOnClickListener {
+            if (isNetworkAvailable(requireContext())) {
+                verifyStockBeforeCheckout()
+            } else {
+                Toast.makeText(context, "\n" +
+                        "Payment cannot be made offline. Please try again later.", Toast.LENGTH_LONG).show()
+            }
+        }
 
         loadCartItems()
-
-        checkoutButton.setOnClickListener {
-            verifyStockBeforeCheckout()
-        }
 
         return view
     }
@@ -88,54 +94,53 @@ class CheckoutFragment : Fragment() {
                         updateTotalPrice()
                         updateItemCount(itemCount)
 
-                        checkoutButton.isEnabled = cartItems.isNotEmpty()
                         updateCartStatus(cartItems.isNotEmpty())
                     }
                 } catch (e: Exception) {
                     withContext(Dispatchers.Main) {
-                        Toast.makeText(context, "Error loading the cart", Toast.LENGTH_SHORT).show()
+                        Toast.makeText(context, "\n" +
+                                "Error loading cart.", Toast.LENGTH_SHORT).show()
                     }
                 }
             }
         }
     }
 
-
-    // Agregar y eliminar productos de forma local
+    // No permitir disminuir la cantidad de productos si no hay conexión
     private fun decreaseCartItemQuantity(cartItem: CartItem) {
+        if (!isNetworkAvailable(requireContext())) {
+            Toast.makeText(context, "You cannot modify the cart offline. Please try again later.", Toast.LENGTH_LONG).show()
+            return
+        }
+
         val db = FirebaseFirestore.getInstance()
         val userId = FirebaseAuth.getInstance().currentUser?.uid
 
         if (userId != null) {
             val cartRef = db.collection("carts").document(userId).collection("items")
 
-            // Si la cantidad es mayor que 1, solo disminuimos la cantidad en el carrito
             if (cartItem.quantity > 1) {
                 cartItem.quantity -= 1
                 cartRef.document(cartItem.id).update("quantity", cartItem.quantity)
                     .addOnSuccessListener {
-                        loadCartItems() // Actualizar la vista del carrito
+                        loadCartItems()
                     }
             } else {
-                // Si es 1, eliminamos el producto del carrito
                 cartRef.document(cartItem.id).delete()
                     .addOnSuccessListener {
-                        loadCartItems() // Recargar los productos del carrito
+                        loadCartItems()
                     }
             }
         }
     }
-    private fun updateCartStatus(hasItems: Boolean) {
-        if (isAdded) {
-            val sharedPreferences = requireContext().getSharedPreferences("CartPrefs", Context.MODE_PRIVATE)
-            val editor = sharedPreferences.edit()
-            editor.putBoolean("hasItemsInCart", hasItems)
-            editor.apply()
-        }
-    }
 
-
+    // No permitir aumentar la cantidad de productos si no hay conexión
     private fun updateCartItemQuantity(cartItem: CartItem) {
+        if (!isNetworkAvailable(requireContext())) {
+            Toast.makeText(context, "You cannot modify the cart offline. Please try again later.", Toast.LENGTH_LONG).show()
+            return
+        }
+
         val db = FirebaseFirestore.getInstance()
         val userId = FirebaseAuth.getInstance().currentUser?.uid
 
@@ -144,12 +149,11 @@ class CheckoutFragment : Fragment() {
                 .document(cartItem.id)
                 .update("quantity", cartItem.quantity)
                 .addOnSuccessListener {
-                    loadCartItems() // Recargar los productos del carrito
+                    loadCartItems()
                 }
         }
     }
 
-    // Verificar stock actual en Firebase al hacer checkout
     private fun verifyStockBeforeCheckout() {
         val db = FirebaseFirestore.getInstance()
         val userId = FirebaseAuth.getInstance().currentUser?.uid
@@ -158,7 +162,7 @@ class CheckoutFragment : Fragment() {
             val cartRef = db.collection("carts").document(userId).collection("items")
 
             GlobalScope.launch(Dispatchers.IO) {
-                val snapshot = cartRef.get().await() // Esperar a que la operación termine sin bloquear el hilo principal
+                val snapshot = cartRef.get().await()
                 var allAvailable = true
 
                 snapshot.documents.forEach { document ->
@@ -174,16 +178,20 @@ class CheckoutFragment : Fragment() {
                                 if (cartItem.quantity > availableQuantity) {
                                     allAvailable = false
                                     withContext(Dispatchers.Main) {
-                                        Toast.makeText(context, "The available quantity of  ${cartItem.productName} is just  $availableQuantity.", Toast.LENGTH_LONG).show()
+                                        if (isAdded) {
+                                            Toast.makeText(requireContext(), "\n" +
+                                                    "The available quantity of ${cartItem.productName} is just $availableQuantity.", Toast.LENGTH_LONG).show()
+                                        }
                                     }
                                 } else {
-                                    // Si todo está bien, restamos la cantidad del producto en Firebase
                                     productRef.update("quantity", availableQuantity - cartItem.quantity).await()
                                 }
                             } else {
                                 allAvailable = false
                                 withContext(Dispatchers.Main) {
-                                    Toast.makeText(context, "Product ${cartItem.productName} not found in inventory.", Toast.LENGTH_LONG).show()
+                                    if (isAdded) {
+                                        Toast.makeText(requireContext(), "Product ${cartItem.productName} not found in inventory.", Toast.LENGTH_LONG).show()
+                                    }
                                 }
                             }
                         }
@@ -192,42 +200,45 @@ class CheckoutFragment : Fragment() {
 
                 if (allAvailable) {
                     withContext(Dispatchers.Main) {
-                        proceedToPayment() // Navegamos al fragmento de pago
+                        if (isAdded) {
+                            proceedToPayment()
+                        }
                     }
                 }
             }
         }
     }
 
-
-
-    // Actualizar el precio total
     private fun updateTotalPrice() {
-        // Calcula el subtotal sin impuestos
         val subtotal = cartAdapter.cartItemList.sumOf { cartItem ->
             val cleanPrice = cartItem.productPrice.replace("$", "").toDoubleOrNull() ?: 0.0
             cleanPrice * cartItem.quantity
         }
 
-        // Calcula los impuestos (7%)
         val taxes = subtotal * 0.07
-
-        // Calcula el total con impuestos
         val totalWithTaxes = subtotal + taxes
 
-        // Actualiza los TextViews correspondientes en la interfaz
         val subtotalTextView = view?.findViewById<TextView>(R.id.subtotal_text_view)
         subtotalTextView?.text = "Subtotal: $${String.format("%.2f", subtotal)}"
 
         val taxesTextView = view?.findViewById<TextView>(R.id.taxes_text_view)
-        taxesTextView?.text = "Taxes (7%): $${String.format("%.2f", taxes)}"
+        taxesTextView?.text = "Impuestos (7%): $${String.format("%.2f", taxes)}"
 
         totalPriceTextView.text = "Total: $${String.format("%.2f", totalWithTaxes)}"
     }
 
     private fun updateItemCount(itemCount: Int) {
         val itemCountTextView = view?.findViewById<TextView>(R.id.item_count_text_view)
-        itemCountTextView?.text = "Total of items: $itemCount"
+        itemCountTextView?.text = "Total de artículos: $itemCount"
+    }
+
+    private fun updateCartStatus(hasItems: Boolean) {
+        if (isAdded) {
+            val sharedPreferences = requireContext().getSharedPreferences("CartPrefs", Context.MODE_PRIVATE)
+            val editor = sharedPreferences.edit()
+            editor.putBoolean("hasItemsInCart", hasItems)
+            editor.apply()
+        }
     }
 
     private fun proceedToPayment() {
@@ -238,9 +249,50 @@ class CheckoutFragment : Fragment() {
             transaction.commit()
         }
     }
+
+    private fun isNetworkAvailable(context: Context): Boolean {
+        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val activeNetwork = connectivityManager.activeNetworkInfo
+        return activeNetwork != null && activeNetwork.isConnected
+    }
+
+    override fun onResume() {
+        super.onResume()
+        if (isNetworkAvailable(requireContext())) {
+            syncCartWithFirebase()
+        }
+    }
+
+    private fun syncCartWithFirebase() {
+        val db = FirebaseFirestore.getInstance()
+        val userId = FirebaseAuth.getInstance().currentUser?.uid
+
+        if (userId != null) {
+            val sharedPreferences = requireContext().getSharedPreferences("CartData", Context.MODE_PRIVATE)
+            val allEntries = sharedPreferences.all
+
+            if (allEntries.isNotEmpty()) {
+                val cartRef = db.collection("carts").document(userId).collection("items")
+
+                GlobalScope.launch(Dispatchers.IO) {
+                    for ((key, value) in allEntries) {
+                        val cartItemId = key
+                        val quantity = value as Int
+
+                        cartRef.document(cartItemId).update("quantity", quantity).await()
+                    }
+
+                    sharedPreferences.edit().clear().apply()
+
+                    withContext(Dispatchers.Main) {
+                        Toast.makeText(context, "\n" +
+                                "Cart synced successfully.", Toast.LENGTH_SHORT).show()
+                    }
+                }
+            }
+        }
+    }
 }
-
-
